requirements.txt
pandas
torch
torch-geometric
spacy
scikit-learn
pyyaml
transformers
networkx
tqdm
flask
flask-cors
nltk
